chore(tests): split backup tests into file per backend
The backup and restore tests were living in the same file for all the
different backend (minio, azurite, azure) therefore we had a big
confusing file, now, every backend has is own file with is own functions
to work with.
