+++ tests/e2e/backup_restore_minio_test.go
@@ -25,26 +25,19 @@ import (
 	apiv1 "github.com/cloudnative-pg/cloudnative-pg/api/v1"
 	"github.com/cloudnative-pg/cloudnative-pg/tests"
 	testUtils "github.com/cloudnative-pg/cloudnative-pg/tests/utils"
+	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/minio"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
 
-var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
+var _ = Describe("MinIO - Backup and restore", Label(tests.LabelBackupRestore), func() {
 	const (
-		level = tests.High
-
-		azuriteBlobSampleFile = fixturesDir + "/backup/azurite/cluster-backup.yaml.template"
-
-		tableName = "to_restore"
-
+		tableName                 = "to_restore"
 		barmanCloudBackupLogEntry = "Starting barman-cloud-backup"
 	)
-
-	currentTimestamp := new(string)
-
 	BeforeEach(func() {
-		if testLevelEnv.Depth < int(level) {
+		if testLevelEnv.Depth < int(tests.High) {
 			Skip("Test depth is lower than the amount requested for this test")
 		}
 	})
@@ -92,7 +85,7 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 				primaryPod, err := env.GetClusterPrimary(namespace, clusterName)
 				Expect(err).ToNot(HaveOccurred())
 				Eventually(func() (bool, error) {
-					connectionStatus, err := testUtils.MinioTestConnectivityUsingBarmanCloudWalArchive(
+					connectionStatus, err := minio.TestConnectivityUsingBarmanCloudWalArchive(
 						namespace, clusterName, primaryPod.GetName(), "minio", "minio123", minioEnv.ServiceName)
 					if err != nil {
 						return false, err
@@ -102,7 +95,7 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 			})
 		})
 
-		// We backup and restore a cluster, and verify some expected data to
+		// We back up and restore a cluster, and verify some expected data to
 		// be there
 		It("backs up and restores a cluster using minio", func() {
 			const (
@@ -137,9 +130,9 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 			// There should be a backup resource and
 			By(fmt.Sprintf("backing up a cluster and verifying it exists on minio, backup path is %v", latestTar), func() {
 				backup = testUtils.ExecuteBackup(namespace, backupFile, false, testTimeouts[testUtils.BackupIsReady], env)
-				AssertBackupConditionInClusterStatus(namespace, clusterName)
+				testUtils.AssertBackupConditionInClusterStatus(env, namespace, clusterName)
 				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnMinio(minioEnv, latestTar)
+					return minio.CountFilesOnMinio(minioEnv, latestTar)
 				}, 60).Should(BeEquivalentTo(1))
 				Eventually(func() (string, error) {
 					cluster, err := env.GetCluster(namespace, clusterName)
@@ -187,7 +180,7 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 
 			By("executing a second backup and verifying the number of backups on minio", func() {
 				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnMinio(minioEnv, latestTar)
+					return minio.CountFilesOnMinio(minioEnv, latestTar)
 				}, 60).Should(BeEquivalentTo(1))
 
 				// delete the first backup and create a second backup
@@ -202,7 +195,7 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 				testUtils.ExecuteBackup(namespace, backupFile, false, testTimeouts[testUtils.BackupIsReady], env)
 				latestTar = minioPath(clusterName, "data.tar")
 				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnMinio(minioEnv, latestTar)
+					return minio.CountFilesOnMinio(minioEnv, latestTar)
 				}, 60).Should(BeEquivalentTo(2))
 			})
 
@@ -233,7 +226,7 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 			latestGZ := filepath.Join("*", clusterName, "*", "*.history.gz")
 			By(fmt.Sprintf("checking the previous number of .history files in minio, history file name is %v",
 				latestGZ), func() {
-				previous, err = testUtils.CountFilesOnMinio(minioEnv, latestGZ)
+				previous, err = minio.CountFilesOnMinio(minioEnv, latestGZ)
 				Expect(err).ToNot(HaveOccurred())
 			})
 
@@ -241,7 +234,7 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 
 			By("checking the number of .history after switchover", func() {
 				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnMinio(minioEnv, latestGZ)
+					return minio.CountFilesOnMinio(minioEnv, latestGZ)
 				}, 60).Should(BeNumerically(">", previous))
 			})
 
@@ -290,9 +283,9 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 			By(fmt.Sprintf("backing up a cluster from standby and verifying it exists on minio, backup path is %v",
 				latestTar), func() {
 				testUtils.ExecuteBackup(namespace, backupStandbyFile, true, testTimeouts[testUtils.BackupIsReady], env)
-				AssertBackupConditionInClusterStatus(namespace, targetClusterName)
+				testUtils.AssertBackupConditionInClusterStatus(env, namespace, targetClusterName)
 				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnMinio(minioEnv, latestTar)
+					return minio.CountFilesOnMinio(minioEnv, latestTar)
 				}, 60).Should(BeEquivalentTo(1))
 				Eventually(func() (string, error) {
 					cluster, err := env.GetCluster(namespace, targetClusterName)
@@ -340,9 +333,9 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 			By(fmt.Sprintf("backing up a cluster from standby (defined in backup file) and verifying it exists on minio,"+
 				" backup path is %v", latestTar), func() {
 				testUtils.ExecuteBackup(namespace, backupWithTargetFile, true, testTimeouts[testUtils.BackupIsReady], env)
-				AssertBackupConditionInClusterStatus(namespace, targetClusterName)
+				testUtils.AssertBackupConditionInClusterStatus(env, namespace, targetClusterName)
 				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnMinio(minioEnv, latestTar)
+					return minio.CountFilesOnMinio(minioEnv, latestTar)
 				}, 60).Should(BeEquivalentTo(1))
 				Eventually(func() (string, error) {
 					cluster, err := env.GetCluster(namespace, targetClusterName)
@@ -357,7 +350,7 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 		})
 
 		// Test that the restore works if the source cluster has a custom
-		// backup.barmanObjectStore.serverName that is different than the cluster name
+		// backup.barmanObjectStore.serverName that is different from the cluster name
 		It("backs up and restores a cluster with custom backup serverName", func() {
 			const (
 				targetDBOne              = "test"
@@ -397,10 +390,10 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 			// There should be a backup resource and
 			By("backing up a cluster and verifying it exists on minio", func() {
 				testUtils.ExecuteBackup(namespace, backupFileCustom, false, testTimeouts[testUtils.BackupIsReady], env)
-				AssertBackupConditionInClusterStatus(namespace, customClusterName)
+				testUtils.AssertBackupConditionInClusterStatus(env, namespace, customClusterName)
 				latestBaseTar := minioPath(clusterServerName, "data.tar")
 				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnMinio(minioEnv, latestBaseTar)
+					return minio.CountFilesOnMinio(minioEnv, latestBaseTar)
 				}, 60).Should(BeEquivalentTo(1),
 					fmt.Sprintf("verify the number of backup %v is equals to 1", latestBaseTar))
 				// this is the second backup we take on the bucket
@@ -436,7 +429,7 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 			// AssertScheduledBackupsImmediate creates at least two backups, we should find
 			// their base backups
 			Eventually(func() (int, error) {
-				return testUtils.CountFilesOnMinio(minioEnv, latestBaseTar)
+				return minio.CountFilesOnMinio(minioEnv, latestBaseTar)
 			}, 60).Should(BeNumerically(">=", 2),
 				fmt.Sprintf("verify the number of backup %v is >= 2", latestBaseTar))
 		})
@@ -446,7 +439,7 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 				restoredClusterName = "restore-cluster-pitr-minio"
 				backupFilePITR      = fixturesDir + "/backup/minio/backup-minio-pitr.yaml"
 			)
-
+			currentTimestamp := new(string)
 			prepareClusterForPITROnMinio(
 				namespace,
 				clusterName,
@@ -487,7 +480,7 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 				AssertScheduledBackupsAreScheduled(namespace, scheduledBackupSampleFile, 300)
 				latestTar := minioPath(clusterName, "data.tar")
 				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnMinio(minioEnv, latestTar)
+					return minio.CountFilesOnMinio(minioEnv, latestTar)
 				}, 60).Should(BeNumerically(">=", 2),
 					fmt.Sprintf("verify the number of backup %v is great than 2", latestTar))
 			})
@@ -497,7 +490,7 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 
 		It("verify tags in backed files", func() {
 			AssertArchiveWalOnMinio(namespace, clusterName, clusterName)
-			tags, err := testUtils.GetFileTagsOnMinio(minioEnv, minioPath(clusterName, "*1.gz"))
+			tags, err := minio.GetFileTagsOnMinio(minioEnv, minioPath(clusterName, "*1.gz"))
 			Expect(err).ToNot(HaveOccurred())
 			Expect(tags.Tags).ToNot(BeEmpty())
 
@@ -513,294 +506,26 @@ var _ = Describe("Backup and restore", Label(tests.LabelBackupRestore), func() {
 
 			AssertNewPrimary(namespace, clusterName, oldPrimary)
 
-			tags, err = testUtils.GetFileTagsOnMinio(minioEnv, minioPath(clusterName, "*.history.gz"))
+			tags, err = minio.GetFileTagsOnMinio(minioEnv, minioPath(clusterName, "*.history.gz"))
 			Expect(err).ToNot(HaveOccurred())
 			Expect(tags.Tags).ToNot(BeEmpty())
 		})
 	})
-
-	Context("using azure blobs as object storage with storage account access authentication", Ordered, func() {
-		// We must be careful here. All the clusters use the same remote storage
-		// and that means that we must use different cluster names otherwise
-		// we risk mixing WALs and backups
-		const azureBlobSampleFile = fixturesDir + "/backup/azure_blob/cluster-with-backup-azure-blob.yaml.template"
-		const clusterRestoreSampleFile = fixturesDir + "/backup/azure_blob/cluster-from-restore.yaml.template"
-		const scheduledBackupSampleFile = fixturesDir +
-			"/backup/scheduled_backup_immediate/scheduled-backup-immediate-azure-blob.yaml"
-		backupFile := fixturesDir + "/backup/azure_blob/backup-azure-blob.yaml"
-		var namespace, clusterName string
-
-		BeforeAll(func() {
-			if !IsAKS() {
-				Skip("This test is only run on AKS clusters")
-			}
-			const namespacePrefix = "cluster-backup-azure-blob"
-			var err error
-			clusterName, err = env.GetResourceNameFromYAML(azureBlobSampleFile)
-			Expect(err).ToNot(HaveOccurred())
-
-			// Create a cluster in a namespace we'll delete after the test
-			namespace, err = env.CreateUniqueTestNamespace(namespacePrefix)
-			Expect(err).ToNot(HaveOccurred())
-
-			// The Azure Blob Storage should have been created ad-hoc for the test.
-			// The credentials are retrieved from the environment variables, as we can't create
-			// a fixture for them
-			By("creating the Azure Blob Storage credentials", func() {
-				AssertStorageCredentialsAreCreated(
-					namespace,
-					"backup-storage-creds",
-					env.AzureConfiguration.StorageAccount,
-					env.AzureConfiguration.StorageKey,
-				)
-			})
-
-			// Create the cluster
-			AssertCreateCluster(namespace, clusterName, azureBlobSampleFile, env)
-		})
-
-		// We backup and restore a cluster, and verify some expected data to
-		// be there
-		It("backs up and restore a cluster", func() {
-			// Write a table and some data on the "app" database
-			tableLocator := TableLocator{
-				Namespace:    namespace,
-				ClusterName:  clusterName,
-				DatabaseName: testUtils.AppDBName,
-				TableName:    tableName,
-			}
-			AssertCreateTestData(env, tableLocator)
-			AssertArchiveWalOnAzureBlob(namespace, clusterName, env.AzureConfiguration)
-			By("uploading a backup", func() {
-				// We create a backup
-				testUtils.ExecuteBackup(namespace, backupFile, false, testTimeouts[testUtils.BackupIsReady], env)
-				AssertBackupConditionInClusterStatus(namespace, clusterName)
-
-				// Verifying file called data.tar should be available on Azure blob storage
-				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnAzureBlobStorage(env.AzureConfiguration, clusterName, "data.tar")
-				}, 30).Should(BeNumerically(">=", 1))
-				Eventually(func() (string, error) {
-					cluster, err := env.GetCluster(namespace, clusterName)
-					return cluster.Status.FirstRecoverabilityPoint, err
-				}, 30).ShouldNot(BeEmpty())
-			})
-
-			// Restore backup in a new cluster
-			AssertClusterRestore(namespace, clusterRestoreSampleFile, tableName)
-
-			By("deleting the restored cluster", func() {
-				err := DeleteResourcesFromFile(namespace, clusterRestoreSampleFile)
-				Expect(err).ToNot(HaveOccurred())
-			})
-		})
-
-		// Create a scheduled backup with the 'immediate' option enabled. We expect the backup to be available
-		It("immediately starts a backup using ScheduledBackups 'immediate' option", func() {
-			scheduledBackupName, err := env.GetResourceNameFromYAML(scheduledBackupSampleFile)
-			Expect(err).ToNot(HaveOccurred())
-
-			AssertScheduledBackupsImmediate(namespace, scheduledBackupSampleFile, scheduledBackupName)
-
-			// Only one data.tar files should be present
-			Eventually(func() (int, error) {
-				return testUtils.CountFilesOnAzureBlobStorage(env.AzureConfiguration,
-					clusterName, "data.tar")
-			}, 30).Should(BeNumerically("==", 2))
-		})
-
-		It("backs up and restore a cluster with PITR", func() {
-			restoredClusterName := "restore-cluster-azure-pitr"
-
-			prepareClusterForPITROnAzureBlob(
-				namespace,
-				clusterName,
-				backupFile,
-				env.AzureConfiguration,
-				2,
-				currentTimestamp,
-			)
-
-			AssertArchiveWalOnAzureBlob(namespace, clusterName, env.AzureConfiguration)
-
-			cluster, err := testUtils.CreateClusterFromBackupUsingPITR(
-				namespace,
-				restoredClusterName,
-				backupFile,
-				*currentTimestamp,
-				env,
-			)
-			Expect(err).ToNot(HaveOccurred())
-			AssertClusterIsReady(namespace, restoredClusterName, testTimeouts[testUtils.ClusterIsReady], env)
-
-			// Restore backup in a new cluster, also cover if no application database is configured
-			AssertClusterWasRestoredWithPITR(namespace, restoredClusterName, tableName, "00000002")
-			By("deleting the restored cluster", func() {
-				Expect(testUtils.DeleteObject(env, cluster)).To(Succeed())
-			})
-		})
-
-		// We create a cluster, create a scheduled backup, patch it to suspend its
-		// execution. We verify that the number of backups does not increase.
-		// We then patch it again back to its initial state and verify that
-		// the amount of backups keeps increasing again
-		It("verifies that scheduled backups can be suspended", func() {
-			const scheduledBackupSampleFile = fixturesDir +
-				"/backup/scheduled_backup_suspend/scheduled-backup-suspend-azure-blob.yaml"
-			scheduledBackupName, err := env.GetResourceNameFromYAML(scheduledBackupSampleFile)
-			Expect(err).ToNot(HaveOccurred())
-
-			By("scheduling backups", func() {
-				AssertScheduledBackupsAreScheduled(namespace, scheduledBackupSampleFile, 480)
-
-				// AssertScheduledBackupsImmediate creates at least two backups, we should find
-				// their base backups
-				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnAzureBlobStorage(env.AzureConfiguration,
-						clusterName, "data.tar")
-				}, 60).Should(BeNumerically(">=", 2))
-			})
-			AssertSuspendScheduleBackups(namespace, scheduledBackupName)
-		})
-	})
-
-	Context("using Azurite blobs as object storage", Ordered, func() {
-		// This is a set of tests using an Azurite server deployed in the same
-		// namespace as the cluster. Since each cluster is installed in its
-		// own namespace, they can share the configuration file
-		const (
-			clusterRestoreSampleFile  = fixturesDir + "/backup/azurite/cluster-from-restore.yaml.template"
-			scheduledBackupSampleFile = fixturesDir +
-				"/backup/scheduled_backup_suspend/scheduled-backup-suspend-azurite.yaml"
-			scheduledBackupImmediateSampleFile = fixturesDir +
-				"/backup/scheduled_backup_immediate/scheduled-backup-immediate-azurite.yaml"
-			backupFile        = fixturesDir + "/backup/azurite/backup.yaml"
-			azuriteCaSecName  = "azurite-ca-secret"
-			azuriteTLSSecName = "azurite-tls-secret"
-		)
-		var namespace, clusterName string
-
-		BeforeAll(func() {
-			if !(IsLocal() || IsGKE() || IsOpenshift()) {
-				Skip("This test is only executed on gke, openshift and local")
-			}
-			const namespacePrefix = "cluster-backup-azurite"
-			var err error
-			clusterName, err = env.GetResourceNameFromYAML(azuriteBlobSampleFile)
-			Expect(err).ToNot(HaveOccurred())
-
-			// Create a cluster in a namespace we'll delete after the test
-			namespace, err = env.CreateUniqueTestNamespace(namespacePrefix)
-			Expect(err).ToNot(HaveOccurred())
-
-			// Create and assert ca and tls certificate secrets on Azurite
-			By("creating ca and tls certificate secrets", func() {
-				err := testUtils.CreateCertificateSecretsOnAzurite(namespace, clusterName,
-					azuriteCaSecName, azuriteTLSSecName, env)
-				Expect(err).ToNot(HaveOccurred())
-			})
-			// Setup Azurite and az cli along with Postgresql cluster
-			prepareClusterBackupOnAzurite(namespace, clusterName, azuriteBlobSampleFile, backupFile, tableName)
-		})
-
-		It("restores a backed up cluster", func() {
-			// Restore backup in a new cluster
-			AssertClusterRestoreWithApplicationDB(namespace, clusterRestoreSampleFile, tableName)
-		})
-
-		// Create a scheduled backup with the 'immediate' option enabled.
-		// We expect the backup to be available
-		It("immediately starts a backup using ScheduledBackups immediate option", func() {
-			scheduledBackupName, err := env.GetResourceNameFromYAML(scheduledBackupImmediateSampleFile)
-			Expect(err).ToNot(HaveOccurred())
-
-			AssertScheduledBackupsImmediate(namespace, scheduledBackupImmediateSampleFile, scheduledBackupName)
-
-			// AssertScheduledBackupsImmediate creates at least two backups, we should find
-			// their base backups
-			Eventually(func() (int, error) {
-				return testUtils.CountFilesOnAzuriteBlobStorage(namespace, clusterName, "data.tar")
-			}, 30).Should(BeNumerically("==", 2))
-		})
-
-		It("backs up and restore a cluster with PITR Azurite", func() {
-			const (
-				restoredClusterName = "restore-cluster-pitr-azurite"
-				backupFilePITR      = fixturesDir + "/backup/azurite/backup-pitr.yaml"
-			)
-
-			prepareClusterForPITROnAzurite(namespace, clusterName, backupFilePITR, currentTimestamp)
-
-			cluster, err := testUtils.CreateClusterFromBackupUsingPITR(
-				namespace,
-				restoredClusterName,
-				backupFilePITR,
-				*currentTimestamp,
-				env,
-			)
-			Expect(err).NotTo(HaveOccurred())
-			AssertClusterIsReady(namespace, restoredClusterName, testTimeouts[testUtils.ClusterIsReady], env)
-
-			// Restore backup in a new cluster, also cover if no application database is configured
-			AssertClusterWasRestoredWithPITR(namespace, restoredClusterName, tableName, "00000002")
-
-			By("deleting the restored cluster", func() {
-				Expect(testUtils.DeleteObject(env, cluster)).To(Succeed())
-			})
-		})
-
-		// We create a cluster, create a scheduled backup, patch it to suspend its
-		// execution. We verify that the number of backups does not increase.
-		// We then patch it again back to its initial state and verify that
-		// the amount of backups keeps increasing again
-		It("verifies that scheduled backups can be suspended", func() {
-			scheduledBackupName, err := env.GetResourceNameFromYAML(scheduledBackupSampleFile)
-			Expect(err).ToNot(HaveOccurred())
-
-			By("scheduling backups", func() {
-				AssertScheduledBackupsAreScheduled(namespace, scheduledBackupSampleFile, 300)
-				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnAzuriteBlobStorage(namespace, clusterName, "data.tar")
-				}, 60).Should(BeNumerically(">=", 3))
-			})
-
-			AssertSuspendScheduleBackups(namespace, scheduledBackupName)
-		})
-	})
 })
 
-var _ = Describe("Clusters Recovery From Barman Object Store", Label(tests.LabelBackupRestore), func() {
+var _ = Describe("MinIO - Clusters Recovery from Barman Object Store", Label(tests.LabelBackupRestore), func() {
 	const (
 		fixturesBackupDir               = fixturesDir + "/backup/recovery_external_clusters/"
-		azuriteBlobSampleFile           = fixturesDir + "/backup/azurite/cluster-backup.yaml.template"
-		externalClusterFileMinio        = fixturesBackupDir + "external-clusters-minio-03.yaml.template"
 		externalClusterFileMinioReplica = fixturesBackupDir + "external-clusters-minio-replica-04.yaml.template"
+		clusterSourceFileMinio          = fixturesBackupDir + "source-cluster-minio-01.yaml.template"
+		externalClusterFileMinio        = fixturesBackupDir + "external-clusters-minio-03.yaml.template"
 		sourceTakeFirstBackupFileMinio  = fixturesBackupDir + "backup-minio-02.yaml"
 		sourceTakeSecondBackupFileMinio = fixturesBackupDir + "backup-minio-03.yaml"
 		sourceTakeThirdBackupFileMinio  = fixturesBackupDir + "backup-minio-04.yaml"
-		clusterSourceFileMinio          = fixturesBackupDir + "source-cluster-minio-01.yaml.template"
-		sourceBackupFileAzure           = fixturesBackupDir + "backup-azure-blob-02.yaml"
-		clusterSourceFileAzure          = fixturesBackupDir + "source-cluster-azure-blob-01.yaml.template"
-		externalClusterFileAzure        = fixturesBackupDir + "external-clusters-azure-blob-03.yaml.template"
-		sourceBackupFileAzurePITR       = fixturesBackupDir + "backup-azure-blob-pitr.yaml"
-		externalClusterFileAzurite      = fixturesBackupDir + "external-clusters-azurite-03.yaml.template"
-		backupFileAzurite               = fixturesBackupDir + "backup-azurite-02.yaml"
 		tableName                       = "to_restore"
-		clusterSourceFileAzureSAS       = fixturesBackupDir + "cluster-with-backup-azure-blob-sas.yaml.template"
-		clusterRestoreFileAzureSAS      = fixturesBackupDir + "cluster-from-restore-sas.yaml.template"
-		sourceBackupFileAzureSAS        = fixturesBackupDir + "backup-azure-blob-sas.yaml"
-		sourceBackupFileAzurePITRSAS    = fixturesBackupDir + "backup-azure-blob-pitr-sas.yaml"
-		level                           = tests.High
-		minioCaSecName                  = "minio-server-ca-secret"
-		minioTLSSecName                 = "minio-server-tls-secret"
-		azuriteCaSecName                = "azurite-ca-secret"
-		azuriteTLSSecName               = "azurite-tls-secret"
 	)
-
-	currentTimestamp := new(string)
-
 	BeforeEach(func() {
-		if testLevelEnv.Depth < int(level) {
+		if testLevelEnv.Depth < int(tests.High) {
 			Skip("Test depth is lower than the amount requested for this test")
 		}
 	})
@@ -836,7 +561,7 @@ var _ = Describe("Clusters Recovery From Barman Object Store", Label(tests.Label
 				primaryPod, err := env.GetClusterPrimary(namespace, clusterName)
 				Expect(err).ToNot(HaveOccurred())
 				Eventually(func() (bool, error) {
-					connectionStatus, err := testUtils.MinioTestConnectivityUsingBarmanCloudWalArchive(
+					connectionStatus, err := minio.TestConnectivityUsingBarmanCloudWalArchive(
 						namespace, clusterName, primaryPod.GetName(), "minio", "minio123", minioEnv.ServiceName)
 					if err != nil {
 						return false, err
@@ -865,7 +590,7 @@ var _ = Describe("Clusters Recovery From Barman Object Store", Label(tests.Label
 			By("backing up a cluster and verifying it exists on minio", func() {
 				testUtils.ExecuteBackup(namespace, sourceTakeFirstBackupFileMinio, false,
 					testTimeouts[testUtils.BackupIsReady], env)
-				AssertBackupConditionInClusterStatus(namespace, clusterName)
+				testUtils.AssertBackupConditionInClusterStatus(env, namespace, clusterName)
 
 				// TODO: this is to force a CHECKPOINT when we run the backup on standby.
 				// This should be better handled inside ExecuteBackup
@@ -873,7 +598,7 @@ var _ = Describe("Clusters Recovery From Barman Object Store", Label(tests.Label
 
 				latestTar := minioPath(clusterName, "data.tar")
 				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnMinio(minioEnv, latestTar)
+					return minio.CountFilesOnMinio(minioEnv, latestTar)
 				}, 60).Should(BeEquivalentTo(1),
 					fmt.Sprintf("verify the number of backup %v is equals to 1", latestTar))
 				Eventually(func() (string, error) {
@@ -907,9 +632,10 @@ var _ = Describe("Clusters Recovery From Barman Object Store", Label(tests.Label
 		It("restores a cluster with 'PITR' from barman object using 'barmanObjectStore' "+
 			" option in 'externalClusters' section", func() {
 			externalClusterRestoreName := "restore-external-cluster-pitr"
+
+			currentTimestamp := new(string)
 			// We have already written 2 rows in test table 'to_restore' in above test now we will take current
 			// timestamp. It will use to restore cluster from source using PITR
-
 			By("getting currentTimestamp", func() {
 				ts, err := testUtils.GetCurrentTimestamp(namespace, clusterName, env)
 				*currentTimestamp = ts
@@ -935,10 +661,10 @@ var _ = Describe("Clusters Recovery From Barman Object Store", Label(tests.Label
 			By("creating second backup and verifying it exists on minio", func() {
 				testUtils.ExecuteBackup(namespace, sourceTakeSecondBackupFileMinio, false,
 					testTimeouts[testUtils.BackupIsReady], env)
-				AssertBackupConditionInClusterStatus(namespace, clusterName)
+				testUtils.AssertBackupConditionInClusterStatus(env, namespace, clusterName)
 				latestTar := minioPath(clusterName, "data.tar")
 				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnMinio(minioEnv, latestTar)
+					return minio.CountFilesOnMinio(minioEnv, latestTar)
 				}, 60).Should(BeEquivalentTo(2),
 					fmt.Sprintf("verify the number of backup %v is equals to 2", latestTar))
 			})
@@ -975,10 +701,10 @@ var _ = Describe("Clusters Recovery From Barman Object Store", Label(tests.Label
 			By("backing up a cluster and verifying it exists on minio", func() {
 				testUtils.ExecuteBackup(namespace, sourceTakeThirdBackupFileMinio, false,
 					testTimeouts[testUtils.BackupIsReady], env)
-				AssertBackupConditionInClusterStatus(namespace, clusterName)
+				testUtils.AssertBackupConditionInClusterStatus(env, namespace, clusterName)
 				latestTar := minioPath(clusterName, "data.tar")
 				Eventually(func() (int, error) {
-					return testUtils.CountFilesOnMinio(minioEnv, latestTar)
+					return minio.CountFilesOnMinio(minioEnv, latestTar)
 				}, 60).Should(BeEquivalentTo(3),
 					fmt.Sprintf("verify the number of backup %v is great than 3", latestTar))
 			})
@@ -992,259 +718,63 @@ var _ = Describe("Clusters Recovery From Barman Object Store", Label(tests.Label
 			)
 		})
 	})
-
-	Context("using azure blobs as object storage", func() {
-		Context("storage account access authentication", Ordered, func() {
-			var namespace, clusterName string
-			BeforeAll(func() {
-				if !IsAKS() {
-					Skip("This test is only executed on AKS clusters")
-				}
-				const namespacePrefix = "recovery-barman-object-azure"
-				var err error
-				clusterName, err = env.GetResourceNameFromYAML(clusterSourceFileAzure)
-				Expect(err).ToNot(HaveOccurred())
-
-				// Create a cluster in a namespace we'll delete after the test
-				namespace, err = env.CreateUniqueTestNamespace(namespacePrefix)
-				Expect(err).ToNot(HaveOccurred())
-
-				// The Azure Blob Storage should have been created ad-hoc for the test.
-				// The credentials are retrieved from the environment variables, as we can't create
-				// a fixture for them
-				By("creating the Azure Blob Storage credentials", func() {
-					AssertStorageCredentialsAreCreated(namespace, "backup-storage-creds",
-						env.AzureConfiguration.StorageAccount, env.AzureConfiguration.StorageKey)
-				})
-
-				// Create the cluster
-				AssertCreateCluster(namespace, clusterName, clusterSourceFileAzure, env)
-			})
-
-			It("restores a cluster from barman object using 'barmanObjectStore' option in 'externalClusters' section", func() {
-				// Write a table and some data on the "app" database
-				tableLocator := TableLocator{
-					Namespace:    namespace,
-					ClusterName:  clusterName,
-					DatabaseName: testUtils.AppDBName,
-					TableName:    tableName,
-				}
-				AssertCreateTestData(env, tableLocator)
-				AssertArchiveWalOnAzureBlob(namespace, clusterName, env.AzureConfiguration)
-
-				By("backing up a cluster and verifying it exists on azure blob storage", func() {
-					// Create the backup
-					testUtils.ExecuteBackup(namespace, sourceBackupFileAzure, false, testTimeouts[testUtils.BackupIsReady], env)
-					AssertBackupConditionInClusterStatus(namespace, clusterName)
-					// Verifying file called data.tar should be available on Azure blob storage
-					Eventually(func() (int, error) {
-						return testUtils.CountFilesOnAzureBlobStorage(env.AzureConfiguration, clusterName, "data.tar")
-					}, 30).Should(BeNumerically(">=", 1))
-				})
-
-				// Restoring cluster using a recovery barman object store, which is defined
-				// in the externalClusters section
-				AssertClusterRestore(namespace, externalClusterFileAzure, tableName)
-			})
-
-			It("restores a cluster with 'PITR' from barman object using "+
-				"'barmanObjectStore' option in 'externalClusters' section", func() {
-				externalClusterName := "external-cluster-azure-pitr"
-
-				prepareClusterForPITROnAzureBlob(
-					namespace,
-					clusterName,
-					sourceBackupFileAzurePITR,
-					env.AzureConfiguration,
-					1,
-					currentTimestamp,
-				)
-
-				restoredCluster, err := testUtils.CreateClusterFromExternalClusterBackupWithPITROnAzure(
-					namespace,
-					externalClusterName,
-					clusterName,
-					*currentTimestamp,
-					"backup-storage-creds",
-					env.AzureConfiguration.StorageAccount,
-					env.AzureConfiguration.BlobContainer,
-					env)
-				Expect(err).ToNot(HaveOccurred())
-
-				// Restoring cluster using a recovery barman object store, which is defined
-				// in the externalClusters section
-				AssertClusterWasRestoredWithPITRAndApplicationDB(
-					namespace,
-					externalClusterName,
-					tableName,
-					"00000002",
-				)
-
-				By("delete restored cluster", func() {
-					Expect(testUtils.DeleteObject(env, restoredCluster)).To(Succeed())
-				})
-			})
-		})
-
-		Context("storage account SAS Token authentication", Ordered, func() {
-			var namespace, clusterName string
-			BeforeAll(func() {
-				if !IsAKS() {
-					Skip("This test is only executed on AKS clusters")
-				}
-				const namespacePrefix = "cluster-backup-azure-blob-sas"
-				var err error
-				clusterName, err = env.GetResourceNameFromYAML(clusterSourceFileAzureSAS)
-				Expect(err).ToNot(HaveOccurred())
-
-				// Create a cluster in a namespace we'll delete after the test
-				namespace, err = env.CreateUniqueTestNamespace(namespacePrefix)
-				Expect(err).ToNot(HaveOccurred())
-
-				// The Azure Blob Storage should have been created ad-hoc for the test,
-				// we get the credentials from the environment variables as we can't create
-				// a fixture for them
-				By("creating the Azure Blob Container SAS Token credentials", func() {
-					AssertCreateSASTokenCredentials(namespace, env.AzureConfiguration.StorageAccount,
-						env.AzureConfiguration.StorageKey)
-				})
-
-				// Create the Cluster
-				AssertCreateCluster(namespace, clusterName, clusterSourceFileAzureSAS, env)
-			})
-
-			It("restores cluster from barman object using 'barmanObjectStore' option in 'externalClusters' section", func() {
-				// Write a table and some data on the "app" database
-				tableLocator := TableLocator{
-					Namespace:    namespace,
-					ClusterName:  clusterName,
-					DatabaseName: testUtils.AppDBName,
-					TableName:    tableName,
-				}
-				AssertCreateTestData(env, tableLocator)
-
-				// Create a WAL on the primary and check if it arrives in the
-				// Azure Blob Storage within a short time
-				AssertArchiveWalOnAzureBlob(namespace, clusterName, env.AzureConfiguration)
-
-				By("backing up a cluster and verifying it exists on azure blob storage", func() {
-					// We create a Backup
-					testUtils.ExecuteBackup(namespace, sourceBackupFileAzureSAS, false, testTimeouts[testUtils.BackupIsReady], env)
-					AssertBackupConditionInClusterStatus(namespace, clusterName)
-					// Verifying file called data.tar should be available on Azure blob storage
-					Eventually(func() (int, error) {
-						return testUtils.CountFilesOnAzureBlobStorage(env.AzureConfiguration, clusterName, "data.tar")
-					}, 30).Should(BeNumerically(">=", 1))
-				})
-
-				// Restore backup in a new cluster
-				AssertClusterRestoreWithApplicationDB(namespace, clusterRestoreFileAzureSAS, tableName)
-			})
-
-			It("restores a cluster with 'PITR' from barman object using "+
-				"'barmanObjectStore' option in 'externalClusters' section", func() {
-				externalClusterName := "external-cluster-azure-pitr"
-
-				prepareClusterForPITROnAzureBlob(
-					namespace,
-					clusterName,
-					sourceBackupFileAzurePITRSAS,
-					env.AzureConfiguration,
-					1,
-					currentTimestamp,
-				)
-
-				restoredCluster, err := testUtils.CreateClusterFromExternalClusterBackupWithPITROnAzure(
-					namespace,
-					externalClusterName,
-					clusterName,
-					*currentTimestamp,
-					"backup-storage-creds-sas",
-					env.AzureConfiguration.StorageAccount,
-					env.AzureConfiguration.BlobContainer,
-					env)
-				Expect(err).ToNot(HaveOccurred())
-
-				// Restoring cluster using a recovery barman object store, which is defined
-				// in the externalClusters section
-				AssertClusterWasRestoredWithPITRAndApplicationDB(
-					namespace,
-					externalClusterName,
-					tableName,
-					"00000002",
-				)
-
-				By("delete restored cluster", func() {
-					Expect(testUtils.DeleteObject(env, restoredCluster)).To(Succeed())
-				})
-			})
-		})
-	})
-
-	Context("using Azurite blobs as object storage", Ordered, func() {
-		var namespace, clusterName string
-		BeforeAll(func() {
-			if IsAKS() {
-				Skip("This test is not run on AKS")
-			}
-			const namespacePrefix = "recovery-barman-object-azurite"
-			var err error
-			clusterName, err = env.GetResourceNameFromYAML(azuriteBlobSampleFile)
-			Expect(err).ToNot(HaveOccurred())
-
-			// Create a cluster in a namespace we'll delete after the test
-			namespace, err = env.CreateUniqueTestNamespace(namespacePrefix)
-			Expect(err).ToNot(HaveOccurred())
-
-			// Create and assert ca and tls certificate secrets on Azurite
-			By("creating ca and tls certificate secrets", func() {
-				err := testUtils.CreateCertificateSecretsOnAzurite(
-					namespace,
-					clusterName,
-					azuriteCaSecName,
-					azuriteTLSSecName,
-					env)
-				Expect(err).ToNot(HaveOccurred())
-			})
-			// Setup Azurite and az cli along with PostgreSQL cluster
-			prepareClusterBackupOnAzurite(
-				namespace,
-				clusterName,
-				azuriteBlobSampleFile,
-				backupFileAzurite,
-				tableName,
-			)
-		})
-
-		It("restore cluster from barman object using 'barmanObjectStore' option in 'externalClusters' section", func() {
-			// Restore backup in a new cluster
-			AssertClusterRestoreWithApplicationDB(namespace, externalClusterFileAzurite, tableName)
-		})
-
-		It("restores a cluster with 'PITR' from barman object using 'barmanObjectStore' "+
-			" option in 'externalClusters' section", func() {
-			const (
-				externalClusterRestoreName = "restore-external-cluster-pitr"
-				backupFileAzuritePITR      = fixturesBackupDir + "backup-azurite-pitr.yaml"
-			)
-
-			prepareClusterForPITROnAzurite(namespace, clusterName, backupFileAzuritePITR, currentTimestamp)
-
-			//  Create a cluster from a particular time using external backup.
-			restoredCluster, err := testUtils.CreateClusterFromExternalClusterBackupWithPITROnAzurite(
-				namespace, externalClusterRestoreName, clusterName, *currentTimestamp, env)
-			Expect(err).NotTo(HaveOccurred())
-
-			AssertClusterWasRestoredWithPITRAndApplicationDB(
-				namespace,
-				externalClusterRestoreName,
-				tableName,
-				"00000002",
-			)
-
-			By("delete restored cluster", func() {
-				Expect(testUtils.DeleteObject(env, restoredCluster)).To(Succeed())
-			})
-		})
-	})
 })
+
+func prepareClusterForPITROnMinio(
+	namespace,
+	clusterName,
+	backupSampleFile string,
+	expectedVal int,
+	currentTimestamp *string,
+) {
+	const tableNamePitr = "for_restore"
+
+	By("backing up a cluster and verifying it exists on minio", func() {
+		testUtils.ExecuteBackup(namespace, backupSampleFile, false, testTimeouts[testUtils.BackupIsReady], env)
+		latestTar := minioPath(clusterName, "data.tar")
+		Eventually(func() (int, error) {
+			return minio.CountFilesOnMinio(minioEnv, latestTar)
+		}, 60).Should(BeNumerically(">=", expectedVal),
+			fmt.Sprintf("verify the number of backups %v is greater than or equal to %v", latestTar,
+				expectedVal))
+		Eventually(func() (string, error) {
+			cluster, err := env.GetCluster(namespace, clusterName)
+			Expect(err).ToNot(HaveOccurred())
+			return cluster.Status.FirstRecoverabilityPoint, err
+		}, 30).ShouldNot(BeEmpty())
+	})
+
+	// Write a table and insert 2 entries on the "app" database
+	tableLocator := TableLocator{
+		Namespace:    namespace,
+		ClusterName:  clusterName,
+		DatabaseName: testUtils.AppDBName,
+		TableName:    tableNamePitr,
+	}
+	AssertCreateTestData(env, tableLocator)
+
+	By("getting currentTimestamp", func() {
+		ts, err := testUtils.GetCurrentTimestamp(namespace, clusterName, env)
+		*currentTimestamp = ts
+		Expect(err).ToNot(HaveOccurred())
+	})
+
+	By(fmt.Sprintf("writing 3rd entry into test table '%v'", tableNamePitr), func() {
+		forward, conn, err := testUtils.ForwardPSQLConnection(
+			env,
+			namespace,
+			clusterName,
+			testUtils.AppDBName,
+			apiv1.ApplicationUserSecretSuffix,
+		)
+		defer func() {
+			forward.Close()
+		}()
+		Expect(err).ToNot(HaveOccurred())
+
+		insertRecordIntoTable(tableNamePitr, 3, conn)
+	})
+	AssertArchiveWalOnMinio(namespace, clusterName, clusterName)
+	AssertArchiveConditionMet(namespace, clusterName, "5m")
+	testUtils.AssertBackupConditionInClusterStatus(env, namespace, clusterName)
+}
